+++ Assets/PenguinQuest/Code/Controllers/LieDownHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47ecfa2bad6a73c409012eced7381d03
+guid: 81356bfedf5336c40973882ce9730bc5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
