+++ Assets/PenguinQuest/Code/Controllers/StandUpHandler.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 47ecfa2bad6a73c409012eced7381d03
+guid: 03575db054bb040478b656105bc59aa5
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
