+++ Assets/PenguinQuest/Input/PlayerControls.cs
@@ -165,7 +165,7 @@ namespace PenguinQuest.Generated
                 {
                     ""name"": ""XAxis"",
                     ""id"": ""77224c1d-cbc7-42cf-b9ea-c39d78e36eb7"",
-                    ""path"": ""1DAxis(whichSideWins=1)"",
+                    ""path"": ""1DAxis"",
                     ""interactions"": """",
                     ""processors"": """",
                     ""groups"": """",
