+++ Assets/PenguinQuest/Input/PlayerControls.inputactions
@@ -141,7 +141,7 @@
                 {
                     "name": "XAxis",
                     "id": "77224c1d-cbc7-42cf-b9ea-c39d78e36eb7",
-                    "path": "1DAxis(whichSideWins=1)",
+                    "path": "1DAxis",
                     "interactions": "",
                     "processors": "",
                     "groups": "",
