+++ Assets/PenguinQuest/Scenes/FeatureSandbox/fsm_test.unity
@@ -4253,7 +4253,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 483879405}
-  m_LocalRotation: {x: 0, y: 0, z: -0.000000026077025, w: 1}
+  m_LocalRotation: {x: 3e-45, y: -3e-45, z: -0.000000026077032, w: 1}
   m_LocalPosition: {x: 2, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -4992,11 +4992,15 @@ GameObject:
   m_Component:
   - component: {fileID: 1429373696}
   - component: {fileID: 1429373700}
-  - component: {fileID: 1429373698}
+  - component: {fileID: 1429373703}
   - component: {fileID: 1429373697}
   - component: {fileID: 1429373699}
   - component: {fileID: 1429373702}
-  - component: {fileID: 1429373703}
+  - component: {fileID: 1429373705}
+  - component: {fileID: 1429373704}
+  - component: {fileID: 1429373706}
+  - component: {fileID: 1429373707}
+  - component: {fileID: 1429373708}
   m_Layer: 9
   m_Name: Player
   m_TagString: Player
@@ -5011,7 +5015,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1429373695}
-  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 3, y: 3, z: 1}
   m_ConstrainProportionsScale: 0
@@ -5040,35 +5044,6 @@ Animator:
   m_HasTransformHierarchy: 1
   m_AllowConstantClipSamplingOptimization: 1
   m_KeepAnimatorControllerStateOnDisable: 0
---- !u!114 &1429373698
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 1429373695}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 166a20cd2a9d4f840ada04bd0fd97a9a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  jumpStrength: 75000
-  jumpAngle: 60
-  locomotionBlendSpeed: 0.1
-  velocityThreshold: 0.01
-  degreesFromSurfaceNormalThreshold: 0.01
-  surfaceAlignmentRotationalStrength: 1
-  maxSpeed: 500
-  enableAutomaticAxisLockingWhenIdle: 1
-  mass: 250
-  centerOfMassX: 0
-  centerOfMassY: 0
-  headCollider: {fileID: 1904906742}
-  torsoCollider: {fileID: 560952285}
-  frontFlipperUpperCollider: {fileID: 1297167880}
-  frontFlipperLowerCollider: {fileID: 1960767493}
-  frontFootCollider: {fileID: 207128432}
-  backFootCollider: {fileID: 1325789281}
 --- !u!114 &1429373699
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -5084,7 +5059,7 @@ MonoBehaviour:
   groundMask:
     serializedVersion: 2
     m_Bits: 256
-  toleratedHeightFromGround: 5
+  toleratedHeightFromGround: 7.5
   displayVisualAids: 1
   rayColorTop: {r: 0, g: 1, b: 1, a: 1}
   rayColorLower: {r: 0, g: 0, b: 1, a: 1}
@@ -5129,14 +5104,11 @@ MonoBehaviour:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1429373695}
-  m_Enabled: 0
+  m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 47ecfa2bad6a73c409012eced7381d03, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  jumpStrength: 50000
-  jumpAngle: 45
-  locomotionBlendSpeed: 0.1
   velocityThreshold: 0.01
   degreesFromSurfaceNormalThreshold: 0.01
   surfaceAlignmentRotationalStrength: 0.1
@@ -5151,7 +5123,84 @@ MonoBehaviour:
   frontFlipperLowerCollider: {fileID: 1960767493}
   frontFootCollider: {fileID: 207128432}
   backFootCollider: {fileID: 1325789281}
-  useInsteadOfOriginal: 0
+--- !u!114 &1429373704
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1429373695}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 81356bfedf5336c40973882ce9730bc5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  headCollider: {fileID: 1904906742}
+  torsoCollider: {fileID: 560952285}
+  frontFlipperUpperCollider: {fileID: 1297167880}
+  frontFlipperLowerCollider: {fileID: 1960767493}
+  frontFootCollider: {fileID: 207128432}
+  backFootCollider: {fileID: 1325789281}
+--- !u!114 &1429373705
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1429373695}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 03575db054bb040478b656105bc59aa5, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  headCollider: {fileID: 1904906742}
+  torsoCollider: {fileID: 560952285}
+  frontFlipperUpperCollider: {fileID: 1297167880}
+  frontFlipperLowerCollider: {fileID: 1960767493}
+  frontFootCollider: {fileID: 207128432}
+  backFootCollider: {fileID: 1325789281}
+--- !u!114 &1429373706
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1429373695}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ffdffde4918d9ad47af7f28a429ad378, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  jumpStrength: 75000
+  jumpAngle: 60
+--- !u!114 &1429373707
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1429373695}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 46117a278e4a3504493788ec6df4f50a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  locomotionBlendStep: 0.1
+--- !u!114 &1429373708
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1429373695}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: d5cc319079e60514385c027aa9bbf717, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  degreesFromSurfaceNormalThreshold: 0.01
+  surfaceAlignmentRotationalStrength: 0.1
+  torsoCollider: {fileID: 560952285}
 --- !u!1001 &1522957906
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -5580,7 +5629,7 @@ Transform:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1788758789}
-  m_LocalRotation: {x: 0, y: 0, z: -0.000000029802312, w: 1}
+  m_LocalRotation: {x: 1e-45, y: 1e-45, z: -0.000000029802312, w: 1}
   m_LocalPosition: {x: 5.22, y: 0.88, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
@@ -5812,7 +5861,7 @@ MonoBehaviour:
     - {x: 0, y: 0, z: 0.13265814, w: 0.9911619}
     - {x: 0, y: 0, z: -0.4310487, w: 0.90232867}
     - {x: 0, y: 0, z: -0.45819503, w: 0.8888517}
-    - {x: 0, y: 0, z: -0.000000029802312, w: 1}
+    - {x: 1e-45, y: 1e-45, z: -0.000000029802312, w: 1}
   m_Iterations: 15
   m_Tolerance: 0.01
   m_Velocity: 0.5
@@ -5997,7 +6046,7 @@ MonoBehaviour:
     m_StoredLocalRotations:
     - {x: 0, y: 0, z: 0.005699533, w: 0.9999838}
     - {x: 0, y: 0, z: -0.037286036, w: 0.99930465}
-    - {x: 0, y: 0, z: -0.000000026077025, w: 1}
+    - {x: 3e-45, y: -3e-45, z: -0.000000026077032, w: 1}
   m_Iterations: 15
   m_Tolerance: 0.01
   m_Velocity: 0.5
